+++ src/ra/utracker.h
@@ -39,6 +39,8 @@
 #ifndef CNC_RED_ALERT_RA_UTRACKER_H_
 #define CNC_RED_ALERT_RA_UTRACKER_H_
 
+#include <cstdint>
+
 /*
 ** UnitTracker Class
 */
@@ -57,14 +59,14 @@ class UnitTrackerClass {
   void Clear_Unit_Total();
 
   int Get_Unit_Total(int unit_type);
-  long* Get_All_Totals();
+  int32_t* Get_All_Totals();
   int Get_Unit_Count() { return UnitCount; }
 
   void To_Network_Format();
   void To_PC_Format();
 
  private:
-  long* UnitTotals;        // Allocate memory for the unit totals
+  int32_t* UnitTotals;     // Allocate memory for the unit totals
   int UnitCount;           // Keep a record of how many unit entries there are
   int InNetworkFormat{0};  // The unit entries are in host format
 };
